lab4

+++ src/main/java/ru/desyatova/trsis/lab2/controller/RentController.java
@@ -6,6 +6,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import ru.desyatova.trsis.lab2.exceptions.ForbiddenException;
 import ru.desyatova.trsis.lab2.services.impl.RentServiceImpl;
 import ru.desyatova.trsis.lab2.entity.Rent;
 import ru.desyatova.trsis.lab2.enums.RentStatus;
@@ -13,6 +14,7 @@ import ru.desyatova.trsis.lab2.exceptions.DuplicateRentException;
 import ru.desyatova.trsis.lab2.exceptions.NotFoundException;
 
 import java.math.BigDecimal;
+import java.security.Principal;
 import java.time.ZonedDateTime;
 import java.util.HashMap;
 import java.util.List;
@@ -30,13 +32,15 @@ public class RentController {
     }
 
     @PostMapping
-    public Rent add(@RequestParam String address, @RequestParam BigDecimal cost, @RequestParam String status) {
+    public Rent add(@RequestParam String address, @RequestParam BigDecimal cost, @RequestParam String status, Principal principal) {
+        if (principal == null) throw new ForbiddenException();
         Rent rent = new Rent(null, address, cost, RentStatus.getByDescription(status));
         return rentServiceImpl.add(rent);
     }
 
     @PutMapping("{id}")
-    public Rent update(@PathVariable Long id, @RequestBody Rent rent) {
+    public Rent update(@PathVariable Long id, @RequestBody Rent rent, Principal principal) {
+        if (principal == null) throw new ForbiddenException();
         if (!id.equals(rent.getId())) throw new IllegalArgumentException("Id в пути не совпадает с id в теле");
         return rentServiceImpl.update(rent);
     }
@@ -47,12 +51,14 @@ public class RentController {
     }
 
     @DeleteMapping("{id}")
-    public void delById(@PathVariable Long id) {
+    public void delById(@PathVariable Long id, Principal principal) {
+        if (principal == null) throw new ForbiddenException();
         rentServiceImpl.delete(id);
     }
 
     @PostMapping("{id}")
-    public Rent addWithID(@PathVariable Long id, @RequestBody Rent rent) {
+    public Rent addWithID(@PathVariable Long id, @RequestBody Rent rent, Principal principal) {
+        if (principal == null) throw new ForbiddenException();
         if (!id.equals(rent.getId())) throw new IllegalArgumentException("Id в пути не совпадает с id в теле");
         return rentServiceImpl.add(rent);
     }
